+++ R/metab.R
@@ -12,13 +12,13 @@
 #'   `model_name` argument to `specs`. The help file for
 #'   [specs()] lists the necessary parameters, describes them in
 #'   detail, and gives default values.
-#' @param data data.frame (not a tbl_df) of input data at the temporal
-#'   resolution of raw observations (unit-value). Columns must have the same
+#' @param data A data frame or tibble of input data at the temporal resolution
+#'   of raw observations (unit-value). Columns must have the same
 #'   names, units, and format as the default. The solar.time column must also
 #'   have a timezone code ('tzone' attribute) of 'UTC'. See the
 #'   **'Formatting `data`'** section below for a full description.
-#' @param data_daily data.frame containing inputs with a daily timestep. See the
-#'   **'Formatting `data_daily`'** section below for a full
+#' @param data_daily A data frame or tibble containing inputs with a daily
+#'   timestep. See the **'Formatting `data_daily`'** section below for a full
 #'   description.
 #' @param info any information, in any format, that you would like to store
 #'   within the metab_model object